# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PixelGlobe API — collaborative pixel canvas
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
}

export interface User {
  id: number;
  username: string;
  pixelCount: number;
  /** @nullable */
  factionId?: number | null;
  /** @nullable */
  factionName?: string | null;
  /**
     * ISO timestamp when cooldown expires
     * @nullable
     */
  cooldownUntil: string | null;
}

export interface Pixel {
  x: number;
  y: number;
  color: string;
  userId: number;
  username: string;
  placedAt: string;
}

export interface CanvasState {
  width: number;
  height: number;
  pixels: Pixel[];
}

export type CanvasStatsTopColorsItem = {
  color: string;
  count: number;
};

export interface CanvasStats {
  totalPixels: number;
  activeUsers: number;
  topColors: CanvasStatsTopColorsItem[];
}

export interface PixelInput {
  x: number;
  y: number;
  color: string;
}

export interface PixelLog {
  x: number;
  y: number;
  color: string;
  username: string;
  /** @nullable */
  factionName?: string | null;
  placedAt: string;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  pixelCount: number;
  /** @nullable */
  factionName?: string | null;
}

export interface UserProfile {
  id: number;
  username: string;
  pixelCount: number;
  /** @nullable */
  factionName?: string | null;
  joinedAt: string;
  recentPixels?: PixelLog[];
}

export interface Faction {
  id: number;
  name: string;
  color: string;
  memberCount: number;
  pixelCount: number;
  /** @nullable */
  description?: string | null;
}

export interface FactionInput {
  /** @minLength 2 */
  name: string;
  color: string;
  description?: string;
}