# lib/api-zod/src/generated/types/canvasState.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PixelGlobe API — collaborative pixel canvas
 * OpenAPI spec version: 0.1.0
 */
import type { Pixel } from './pixel';

export interface CanvasState {
  width: number;
  height: number;
  pixels: Pixel[];
}