# lib/api-zod/src/generated/types/canvasStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PixelGlobe API — collaborative pixel canvas
 * OpenAPI spec version: 0.1.0
 */
import type { CanvasStatsTopColorsItem } from './canvasStatsTopColorsItem';

export interface CanvasStats {
  totalPixels: number;
  activeUsers: number;
  topColors: CanvasStatsTopColorsItem[];
}