# lib/api-zod/src/generated/types/factionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PixelGlobe API — collaborative pixel canvas
 * OpenAPI spec version: 0.1.0
 */

export interface FactionInput {
  /** @minLength 2 */
  name: string;
  color: string;
  description?: string;
}