# lib/api-zod/src/generated/types/pixelLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PixelGlobe API — collaborative pixel canvas
 * OpenAPI spec version: 0.1.0
 */

export interface PixelLog {
  x: number;
  y: number;
  color: string;
  username: string;
  /** @nullable */
  factionName?: string | null;
  placedAt: string;
}