# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PixelGlobe API — collaborative pixel canvas
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  pixelCount: number;
  /** @nullable */
  factionId?: number | null;
  /** @nullable */
  factionName?: string | null;
  /**
     * ISO timestamp when cooldown expires
     * @nullable
     */
  cooldownUntil: string | null;
}