# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PixelGlobe API — collaborative pixel canvas
 * OpenAPI spec version: 0.1.0
 */
import type { PixelLog } from './pixelLog';

export interface UserProfile {
  id: number;
  username: string;
  pixelCount: number;
  /** @nullable */
  factionName?: string | null;
  joinedAt: string;
  recentPixels?: PixelLog[];
}